Int is not enough for 1Tb of data

+++ src/main/scala/TpchSchemaProvider.scala
@@ -4,20 +4,20 @@ import org.apache.spark.SparkContext
 
 // TPC-H table schemas
 case class Customer(
-  c_custkey: Int,
+  c_custkey: Long,
   c_name: String,
   c_address: String,
-  c_nationkey: Int,
+  c_nationkey: Long,
   c_phone: String,
   c_acctbal: Double,
   c_mktsegment: String,
   c_comment: String)
 
 case class Lineitem(
-  l_orderkey: Int,
-  l_partkey: Int,
-  l_suppkey: Int,
-  l_linenumber: Int,
+  l_orderkey: Long,
+  l_partkey: Long,
+  l_suppkey: Long,
+  l_linenumber: Long,
   l_quantity: Double,
   l_extendedprice: Double,
   l_discount: Double,
@@ -32,50 +32,50 @@ case class Lineitem(
   l_comment: String)
 
 case class Nation(
-  n_nationkey: Int,
+  n_nationkey: Long,
   n_name: String,
-  n_regionkey: Int,
+  n_regionkey: Long,
   n_comment: String)
 
 case class Order(
-  o_orderkey: Int,
-  o_custkey: Int,
+  o_orderkey: Long,
+  o_custkey: Long,
   o_orderstatus: String,
   o_totalprice: Double,
   o_orderdate: String,
   o_orderpriority: String,
   o_clerk: String,
-  o_shippriority: Int,
+  o_shippriority: Long,
   o_comment: String)
 
 case class Part(
-  p_partkey: Int,
+  p_partkey: Long,
   p_name: String,
   p_mfgr: String,
   p_brand: String,
   p_type: String,
-  p_size: Int,
+  p_size: Long,
   p_container: String,
   p_retailprice: Double,
   p_comment: String)
 
 case class Partsupp(
-  ps_partkey: Int,
-  ps_suppkey: Int,
-  ps_availqty: Int,
+  ps_partkey: Long,
+  ps_suppkey: Long,
+  ps_availqty: Long,
   ps_supplycost: Double,
   ps_comment: String)
 
 case class Region(
-  r_regionkey: Int,
+  r_regionkey: Long,
   r_name: String,
   r_comment: String)
 
 case class Supplier(
-  s_suppkey: Int,
+  s_suppkey: Long,
   s_name: String,
   s_address: String,
-  s_nationkey: Int,
+  s_nationkey: Long,
   s_phone: String,
   s_acctbal: Double,
   s_comment: String)
@@ -88,28 +88,28 @@ class TpchSchemaProvider(sc: SparkContext, inputDir: String) {
 
   val dfMap = Map(
     "customer" -> sc.textFile(inputDir + "/customer.tbl*").map(_.split('|')).map(p =>
-      Customer(p(0).trim.toInt, p(1).trim, p(2).trim, p(3).trim.toInt, p(4).trim, p(5).trim.toDouble, p(6).trim, p(7).trim)).toDF(),
+      Customer(p(0).trim.toLong, p(1).trim, p(2).trim, p(3).trim.toLong, p(4).trim, p(5).trim.toDouble, p(6).trim, p(7).trim)).toDF(),
 
     "lineitem" -> sc.textFile(inputDir + "/lineitem.tbl*").map(_.split('|')).map(p =>
-      Lineitem(p(0).trim.toInt, p(1).trim.toInt, p(2).trim.toInt, p(3).trim.toInt, p(4).trim.toDouble, p(5).trim.toDouble, p(6).trim.toDouble, p(7).trim.toDouble, p(8).trim, p(9).trim, p(10).trim, p(11).trim, p(12).trim, p(13).trim, p(14).trim, p(15).trim)).toDF(),
+      Lineitem(p(0).trim.toLong, p(1).trim.toLong, p(2).trim.toLong, p(3).trim.toLong, p(4).trim.toDouble, p(5).trim.toDouble, p(6).trim.toDouble, p(7).trim.toDouble, p(8).trim, p(9).trim, p(10).trim, p(11).trim, p(12).trim, p(13).trim, p(14).trim, p(15).trim)).toDF(),
 
     "nation" -> sc.textFile(inputDir + "/nation.tbl*").map(_.split('|')).map(p =>
-      Nation(p(0).trim.toInt, p(1).trim, p(2).trim.toInt, p(3).trim)).toDF(),
+      Nation(p(0).trim.toLong, p(1).trim, p(2).trim.toLong, p(3).trim)).toDF(),
 
     "region" -> sc.textFile(inputDir + "/region.tbl*").map(_.split('|')).map(p =>
-      Region(p(0).trim.toInt, p(1).trim, p(2).trim)).toDF(),
+      Region(p(0).trim.toLong, p(1).trim, p(2).trim)).toDF(),
 
     "order" -> sc.textFile(inputDir + "/orders.tbl*").map(_.split('|')).map(p =>
-      Order(p(0).trim.toInt, p(1).trim.toInt, p(2).trim, p(3).trim.toDouble, p(4).trim, p(5).trim, p(6).trim, p(7).trim.toInt, p(8).trim)).toDF(),
+      Order(p(0).trim.toLong, p(1).trim.toLong, p(2).trim, p(3).trim.toDouble, p(4).trim, p(5).trim, p(6).trim, p(7).trim.toLong, p(8).trim)).toDF(),
 
     "part" -> sc.textFile(inputDir + "/part.tbl*").map(_.split('|')).map(p =>
-      Part(p(0).trim.toInt, p(1).trim, p(2).trim, p(3).trim, p(4).trim, p(5).trim.toInt, p(6).trim, p(7).trim.toDouble, p(8).trim)).toDF(),
+      Part(p(0).trim.toLong, p(1).trim, p(2).trim, p(3).trim, p(4).trim, p(5).trim.toLong, p(6).trim, p(7).trim.toDouble, p(8).trim)).toDF(),
 
     "partsupp" -> sc.textFile(inputDir + "/partsupp.tbl*").map(_.split('|')).map(p =>
-      Partsupp(p(0).trim.toInt, p(1).trim.toInt, p(2).trim.toInt, p(3).trim.toDouble, p(4).trim)).toDF(),
+      Partsupp(p(0).trim.toLong, p(1).trim.toLong, p(2).trim.toLong, p(3).trim.toDouble, p(4).trim)).toDF(),
 
     "supplier" -> sc.textFile(inputDir + "/supplier.tbl*").map(_.split('|')).map(p =>
-      Supplier(p(0).trim.toInt, p(1).trim, p(2).trim, p(3).trim.toInt, p(4).trim, p(5).trim.toDouble, p(6).trim)).toDF())
+      Supplier(p(0).trim.toLong, p(1).trim, p(2).trim, p(3).trim.toLong, p(4).trim, p(5).trim.toDouble, p(6).trim)).toDF())
 
   // for implicits
   val customer = dfMap.get("customer").get
